1.11.0
- Merge all class in "model" package to "AccountInformation" class.
- Moving "Account" class functions to "Accounts" class and this will return each data type in "AccountInformation" class.
- Change some properties in "AccountInfrormation" class (merged from whole class in "model" package). You'll need to modify your code to working again.
- Optimize codes.
