--- src/main/java/io/dutwrapper/dutwrapper/model/accounts/AccountInformation.java
@@ -1,329 +1,0 @@
-package io.dutwrapper.dutwrapper.model.accounts;
-
-import java.io.Serializable;
-
-public class AccountInformation implements Serializable {
-    private String name;
-    private String dateOfBirth;
-    private String birthPlace;
-    private String gender;
-    private String ethnicity;
-    private String nationality;
-    private String nationalIdCard;
-    private String nationalIdCardIssueDate;
-    private String nationalIdCardIssuePlace;
-    private String citizenIdCard;
-    private String citizenIdCardIssueDate;
-    private String religion;
-    private String accountBankId;
-    private String accountBankName;
-    private String hIId;
-    private String hIExpireDate;
-    private String specialization;
-    private String schoolClass;
-    private String trainingProgramPlan;
-    private String trainingProgramPlan2;
-    private String schoolEmail;
-    private String personalEmail;
-    private String schoolEmailInitPass;
-    private String facebookUrl;
-    private String phoneNumber;
-    private String address;
-    private String addressFrom;
-    private String addressCity;
-    private String addressDistrict;
-    private String addressSubDistrict;
-    private String studentId;
-
-    public AccountInformation() {
-
-    }
-
-    public AccountInformation(String name, String dateOfBirth, String birthPlace, String gender, String ethnicity,
-            String nationality, String nationalIdCard, String nationalIdCardIssueDate, String nationalIdCardIssuePlace,
-            String citizenIdCard, String citizenIdCardIssueDate, String religion, String accountBankId,
-            String accountBankName, String hIId, String hIExpireDate, String specialization, String schoolClass,
-            String trainingProgramPlan, String trainingProgramPlan2, String schoolEmail, String personalEmail,
-            String schoolEmailInitPass, String facebookUrl, String phoneNumber, String address, String addressFrom,
-            String addressCity, String addressDistrict, String addressSubDistrict, String studentId) {
-        this.name = name;
-        this.dateOfBirth = dateOfBirth;
-        this.birthPlace = birthPlace;
-        this.gender = gender;
-        this.ethnicity = ethnicity;
-        this.nationality = nationality;
-        this.nationalIdCard = nationalIdCard;
-        this.nationalIdCardIssueDate = nationalIdCardIssueDate;
-        this.nationalIdCardIssuePlace = nationalIdCardIssuePlace;
-        this.citizenIdCard = citizenIdCard;
-        this.citizenIdCardIssueDate = citizenIdCardIssueDate;
-        this.religion = religion;
-        this.accountBankId = accountBankId;
-        this.accountBankName = accountBankName;
-        this.hIId = hIId;
-        this.hIExpireDate = hIExpireDate;
-        this.specialization = specialization;
-        this.schoolClass = schoolClass;
-        this.trainingProgramPlan = trainingProgramPlan;
-        this.trainingProgramPlan2 = trainingProgramPlan2;
-        this.schoolEmail = schoolEmail;
-        this.personalEmail = personalEmail;
-        this.schoolEmailInitPass = schoolEmailInitPass;
-        this.facebookUrl = facebookUrl;
-        this.phoneNumber = phoneNumber;
-        this.address = address;
-        this.addressFrom = addressFrom;
-        this.addressCity = addressCity;
-        this.addressDistrict = addressDistrict;
-        this.addressSubDistrict = addressSubDistrict;
-        this.studentId = studentId;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDateOfBirth() {
-        return dateOfBirth;
-    }
-
-    public void setDateOfBirth(String dateOfBirth) {
-        this.dateOfBirth = dateOfBirth;
-    }
-
-    public String getBirthPlace() {
-        return birthPlace;
-    }
-
-    public void setBirthPlace(String birthPlace) {
-        this.birthPlace = birthPlace;
-    }
-
-    public String getGender() {
-        return gender;
-    }
-
-    public void setGender(String gender) {
-        this.gender = gender;
-    }
-
-    public String getEthnicity() {
-        return ethnicity;
-    }
-
-    public void setEthnicity(String ethnicity) {
-        this.ethnicity = ethnicity;
-    }
-
-    public String getNationality() {
-        return nationality;
-    }
-
-    public void setNationality(String nationality) {
-        this.nationality = nationality;
-    }
-
-    public String getNationalIdCard() {
-        return nationalIdCard;
-    }
-
-    public void setNationalIdCard(String nationalIdCard) {
-        this.nationalIdCard = nationalIdCard;
-    }
-
-    public String getNationalIdCardIssueDate() {
-        return nationalIdCardIssueDate;
-    }
-
-    public void setNationalIdCardIssueDate(String nationalIdCardIssueDate) {
-        this.nationalIdCardIssueDate = nationalIdCardIssueDate;
-    }
-
-    public String getNationalIdCardIssuePlace() {
-        return nationalIdCardIssuePlace;
-    }
-
-    public void setNationalIdCardIssuePlace(String nationalIdCardIssuePlace) {
-        this.nationalIdCardIssuePlace = nationalIdCardIssuePlace;
-    }
-
-    public String getCitizenIdCard() {
-        return citizenIdCard;
-    }
-
-    public void setCitizenIdCard(String citizenIdCard) {
-        this.citizenIdCard = citizenIdCard;
-    }
-
-    public String getCitizenIdCardIssueDate() {
-        return citizenIdCardIssueDate;
-    }
-
-    public void setCitizenIdCardIssueDate(String citizenIdCardIssueDate) {
-        this.citizenIdCardIssueDate = citizenIdCardIssueDate;
-    }
-
-    public String getReligion() {
-        return religion;
-    }
-
-    public void setReligion(String religion) {
-        this.religion = religion;
-    }
-
-    public String getAccountBankId() {
-        return accountBankId;
-    }
-
-    public void setAccountBankId(String accountBankId) {
-        this.accountBankId = accountBankId;
-    }
-
-    public String getAccountBankName() {
-        return accountBankName;
-    }
-
-    public void setAccountBankName(String accountBankName) {
-        this.accountBankName = accountBankName;
-    }
-
-    public String getHIId() {
-        return hIId;
-    }
-
-    public void setHIId(String hIId) {
-        this.hIId = hIId;
-    }
-
-    public String getHIExpireDate() {
-        return hIExpireDate;
-    }
-
-    public void setHIExpireDate(String hIExpireDate) {
-        this.hIExpireDate = hIExpireDate;
-    }
-
-    public String getSpecialization() {
-        return specialization;
-    }
-
-    public void setSpecialization(String specialization) {
-        this.specialization = specialization;
-    }
-
-    public String getSchoolClass() {
-        return schoolClass;
-    }
-
-    public void setSchoolClass(String schoolClass) {
-        this.schoolClass = schoolClass;
-    }
-
-    public String getTrainingProgramPlan() {
-        return trainingProgramPlan;
-    }
-
-    public void setTrainingProgramPlan(String trainingProgramPlan) {
-        this.trainingProgramPlan = trainingProgramPlan;
-    }
-
-    public String getTrainingProgramPlan2() {
-        return trainingProgramPlan2;
-    }
-
-    public void setTrainingProgramPlan2(String trainingProgramPlan2) {
-        this.trainingProgramPlan2 = trainingProgramPlan2;
-    }
-
-    public String getSchoolEmail() {
-        return schoolEmail;
-    }
-
-    public void setSchoolEmail(String schoolEmail) {
-        this.schoolEmail = schoolEmail;
-    }
-
-    public String getPersonalEmail() {
-        return personalEmail;
-    }
-
-    public void setPersonalEmail(String personalEmail) {
-        this.personalEmail = personalEmail;
-    }
-
-    public String getSchoolEmailInitPass() {
-        return schoolEmailInitPass;
-    }
-
-    public void setSchoolEmailInitPass(String schoolEmailInitPass) {
-        this.schoolEmailInitPass = schoolEmailInitPass;
-    }
-
-    public String getFacebookUrl() {
-        return facebookUrl;
-    }
-
-    public void setFacebookUrl(String facebookUrl) {
-        this.facebookUrl = facebookUrl;
-    }
-
-    public String getPhoneNumber() {
-        return phoneNumber;
-    }
-
-    public void setPhoneNumber(String phoneNumber) {
-        this.phoneNumber = phoneNumber;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public String getAddressFrom() {
-        return addressFrom;
-    }
-
-    public void setAddressFrom(String addressFrom) {
-        this.addressFrom = addressFrom;
-    }
-
-    public String getAddressCity() {
-        return addressCity;
-    }
-
-    public void setAddressCity(String addressCity) {
-        this.addressCity = addressCity;
-    }
-
-    public String getAddressDistrict() {
-        return addressDistrict;
-    }
-
-    public void setAddressDistrict(String addressDistrict) {
-        this.addressDistrict = addressDistrict;
-    }
-
-    public String getAddressSubDistrict() {
-        return addressSubDistrict;
-    }
-
-    public void setAddressSubDistrict(String addressSubDistrict) {
-        this.addressSubDistrict = addressSubDistrict;
-    }
-
-    public String getStudentId() {
-        return studentId;
-    }
-
-    public void setStudentId(String studentId) {
-        this.studentId = studentId;
-    }
-}
