--- src/main/java/io/dutwrapper/dutwrapper/model/accounts/WeekItem.java
@@ -1,29 +1,0 @@
-package io.dutwrapper.dutwrapper.model.accounts;
-
-import java.io.Serializable;
-
-public class WeekItem implements Serializable {
-    private Integer start;
-    private Integer end;
-
-    public WeekItem(Integer start, Integer end) {
-        this.start = start;
-        this.end = end;
-    }
-
-    public Integer getStart() {
-        return start;
-    }
-
-    public void setStart(Integer start) {
-        this.start = start;
-    }
-
-    public Integer getEnd() {
-        return end;
-    }
-
-    public void setEnd(Integer end) {
-        this.end = end;
-    }
-}
