+++ src/test/java/io/dutwrapper/dutwrapper/AccountTest.java
@@ -11,6 +11,7 @@ public class AccountTest {
         Gson gson = new GsonBuilder().setPrettyPrinting().create();
 
         String envDutAccount = System.getenv("dut_account");
+        // System.out.println(envDutAccount);
         if (envDutAccount == null)
             throw new Exception("No account found! Please define \"dut_account\" variable to test account library.");
         if (envDutAccount.split("\\|").length != 2)
@@ -20,25 +21,25 @@ public class AccountTest {
         Integer year = 20;
         Integer semester = 2;
 
-        Account.Session session = Account.getSession();
-        Account.AuthInfo authInfo = new Account.AuthInfo(user, pass);
+        Accounts.Session session = Accounts.getSession();
+        Accounts.AuthInfo authInfo = new Accounts.AuthInfo(user, pass);
 
-        System.out.println("Create new Session");
+        System.out.println("Create new session");
         System.out.println(gson.toJson(session));
-        System.out.println(gson.toJson(Account.isLoggedIn(session)));
+        System.out.println(gson.toJson(Accounts.isLoggedIn(session)));
         System.out.println("\nLogin and check again");
-        Account.login(session, authInfo);
-        System.out.println(gson.toJson(Account.isLoggedIn(session)));
+        Accounts.login(session, authInfo);
+        System.out.println(gson.toJson(Accounts.isLoggedIn(session)));
         System.out.println("\nFetch subject schedule list");
-        System.out.println(gson.toJson(Account.fetchSubjectSchedule(session, year, semester)));
+        System.out.println(gson.toJson(Accounts.fetchSubjectSchedule(session, year, semester)));
         System.out.println("\nFetch subject fee list");
-        System.out.println(gson.toJson(Account.fetchSubjectFee(session, year, semester)));
+        System.out.println(gson.toJson(Accounts.fetchSubjectFee(session, year, semester)));
         System.out.println("\nFetch account information");
-        System.out.println(gson.toJson(Account.fetchAccountInformation(session)));
+        System.out.println(gson.toJson(Accounts.fetchStudentInformation(session)));
         System.out.println("\nFetch account training status");
-        System.out.println(gson.toJson(Account.fetchAccountTrainingStatus(session)));
+        System.out.println(gson.toJson(Accounts.fetchTrainingStatus(session)));
         System.out.println("\nLogout and check if logged out");
-        Account.logout(session);
-        System.out.println(gson.toJson(Account.isLoggedIn(session)));
+        Accounts.logout(session);
+        System.out.println(gson.toJson(Accounts.isLoggedIn(session)));
     }
 }
